main screens refactor
