+++ app/lib/views/main screen/orders_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class OrdersScreen extends StatelessWidget {
+  const OrdersScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(child: Text('Заказы'));
+  }
+}
