+++ app/lib/views/main screen/products_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class ProductsScreen extends StatelessWidget {
+  const ProductsScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(child: Text('Менюшка'));
+  }
+}
